split hyper runtime into bin + lib
